change the data type of id from long to integer of the publication repository class

+++ src/main/java/com/system/blog/repositories/PublicationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PublicationRepository extends JpaRepository<Publication,Long> {
+public interface PublicationRepository extends JpaRepository<Publication,Integer> {
 }
